+++ src/Services/Placement/PlacementServiceBuilder.php
@@ -6,6 +6,7 @@ namespace Bitrix24\SDK\Services\Placement;
 
 use Bitrix24\SDK\Services\AbstractServiceBuilder;
 use Bitrix24\SDK\Services\Placement\Service\Placement;
+use Bitrix24\SDK\Services\Placement\Service\UserFieldType;
 
 class PlacementServiceBuilder extends AbstractServiceBuilder
 {
@@ -16,6 +17,18 @@ class PlacementServiceBuilder extends AbstractServiceBuilder
     {
         if (!isset($this->serviceCache[__METHOD__])) {
             $this->serviceCache[__METHOD__] = new Placement($this->core, $this->log);
+        }
+
+        return $this->serviceCache[__METHOD__];
+    }
+
+    /**
+     * @return \Bitrix24\SDK\Services\Placement\Service\UserFieldType
+     */
+    public function userfieldtype(): UserFieldType
+    {
+        if (!isset($this->serviceCache[__METHOD__])) {
+            $this->serviceCache[__METHOD__] = new UserFieldType($this->core, $this->log);
         }
 
         return $this->serviceCache[__METHOD__];